-- fixed a bug inside IOBuffer. Now it properly moves data if space available -- re-implemented timestamp computations for inbound RTSP/RTCP/RTP. Now RTCP packets are properly processed
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@379 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/protocols/rtp/rtpheader.h
@@ -34,7 +34,7 @@
 
 typedef struct DLLEXP _RTPHeader {
 	uint32_t _flags;
-	uint64_t _timestamp;
+	uint32_t _timestamp;
 	uint32_t _ssrc;
 	//we are not interested in this yet
 	//vector<uint32_t> _csrc;
